Intrainz Web Intern Project
JavaScript Form Validation Project

Forms are an integral part of our Website. We need to verify all the details so we know the person filling the form is Human and not a Bot. Create a Form with following fields and check if each fields have relevant data in it.

Required Fields:
 Full Name, Email ID, Phone Number, Password and Confirm Password

Validation Criteria:
 Name must not be less than 5 characters
 Email Id should have @ character in it
 Phone Number should not be 123456789 and must be a 10-digit number.
 Password cannot be ‘password’ or ‘name of the user’ or less than 8 characters.
 Password and confirm password should match
 Whenever any of the above are not met, the error should pop saying ex; “enter correct email” or “password is not strong” etc
 You can include more conditions and fields to the form validation but the above are mandatory.

Must use the following for development:
 HTML
 CSS and Bootstrap
 JavaScript( Event listeners(onSubmit and onChange), multiple functions, loops, if/else )
 Use comments wherever possible to make it readable for others.
 Deploy your code to GitHub and merge with Netlify to make it live.
 Do not use any external library for this.
 Build your own logic do not copy paste from google.
 Design can be as per your vision and make it as creative as you can.

The Project will be judged on the Functionality, if all the mentioned things are used or
not, how clean the code is and lastly the design.